We now need a JSON serlalization object, but the huge numbers we use mostly now work. DB schema modded so  fits in. Lib versions update to more recent

+++ migrations/2018-06-28-132749_create_blocks/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE blocks (
        hash VARCHAR(55),
        height BIGINT,
        miner VARCHAR(55),
-       nonce BIGINT,
+       nonce numeric(20,0),
        prev_hash VARCHAR(55),
        state_hash VARCHAR(55),
        target BIGINT,
